CommsIO/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "commlinkio"
version = "0.1.0"
description = "Communication triggers and integrations for Telegram, Gmail, and more"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "python-telegram-bot>=22.0,<23.0",
  "pydantic>=2.0,<3.0",
  "python-dotenv>=1.0,<2.0",
]

[project.scripts]
commlinkio-telegram-updates-trigger = "commlinkio.telegram.triggers.updates_trigger:main"
commlinkio-gmail-inbox-watch-trigger = "commlinkio.gmail.triggers.inbox_watch_trigger:main"
commlinkio-json-extract = "commlinkio.tools.json_extract:main"
commlinkio-telegram-send = "commlinkio.telegram.tools.send:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
